+++ markwon-core/src/main/java/io/noties/markwon/MarkwonSpansFactory.java
@@ -39,9 +39,33 @@ public interface MarkwonSpansFactory {
          * {@link SpanFactory} with the specified one.
          *
          * @since 3.0.1
+         * @deprecated 4.2.2 consider using {@link #appendFactory(Class, SpanFactory)} or
+         * {@link #prependFactory(Class, SpanFactory)} methods for more explicit factory ordering.
+         * `addFactory` behaved like {@link #prependFactory(Class, SpanFactory)}, so
+         * this method call can be replaced with it
          */
         @NonNull
+        @Deprecated
         <N extends Node> Builder addFactory(@NonNull Class<N> node, @NonNull SpanFactory factory);
+
+        /**
+         * Append a factory to existing one (or make the first one for specified node). Specified factory
+         * will be called <strong>after</strong> original (if present) factory. Can be used to
+         * <em>change</em> behavior or original span factory.
+         *
+         * @since 4.2.2
+         */
+        @NonNull
+        <N extends Node> Builder appendFactory(@NonNull Class<N> node, @NonNull SpanFactory factory);
+
+        /**
+         * Prepend a factory to existing one (or make the first one for specified node). Specified factory
+         * will be called <string>before</string> original (if present) factory.
+         *
+         * @since 4.2.2
+         */
+        @NonNull
+        <N extends Node> Builder prependFactory(@NonNull Class<N> node, @NonNull SpanFactory factory);
 
         /**
          * Can be useful when <em>enhancing</em> an already defined SpanFactory with another one.